Updated pdb.py to use the new pypinproc methods
I changed the names of the methods for the PD-LED board in pypinproc to
be more pythonic. This pdb.py update uses those new names. It also adds
support for brightness compensation for the LEDs. Plus a few other
cleanup items in the code around comments, line breaks, etc. Also
removed the LED ‘script’ functionality since that’s now handled with
the new LEDcontroller.